backend/requirements.txt
uvicorn==0.34.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
greenlet==3.1.1
idna==3.10
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
mysql-connector-python==9.1.0
redis==5.2.1
rapidfuzz==3.11.0
